refactor: fix an error with superjson version
