+++ apps/bot/package.json
@@ -23,7 +23,7 @@
     "@trpc/client": "^10.45.2",
     "discord-api-types": "^0.37.101",
     "discord.js": "^14.7.1",
-    "superjson": "2.2.1"
+    "superjson": "1.9.1"
   },
   "devDependencies": {
     "@discord-bot/eslint-config": "workspace:*",
