+++ packages/api/package.json
@@ -15,7 +15,7 @@
     "@discord-bot/db": "workspace:*",
     "@trpc/client": "^10.45.2",
     "@trpc/server": "^10.45.2",
-    "superjson": "2.2.1",
+    "superjson": "1.9.1",
     "zod": "^3.23.8"
   },
   "devDependencies": {
